+++ .eslintrc.js
@@ -24,6 +24,15 @@ module.exports = {
       './tsconfig.eslint.json',
       './packages/*/tsconfig.json',
       './tests/integration/tsconfig.json',
+      /**
+       * We are currently in the process of transitioning to nx's out of the box structure and
+       * so need to manually specify converted packages' tsconfig.build.json and tsconfig.spec.json
+       * files here for now in addition to the tsconfig.json glob pattern.
+       *
+       * TODO(#4665): Clean this up once all packages have been transitioned.
+       */
+      './packages/scope-manager/tsconfig.build.json',
+      './packages/scope-manager/tsconfig.spec.json',
     ],
     allowAutomaticSingleRunInference: true,
     tsconfigRootDir: __dirname,